Revert "Convert feed files to json, part 221"
