+++ tests/vds/largedocuments.rb
@@ -1,5 +1,4 @@
 # Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-require 'document_set'
 require 'vds_test'
 
 class LargeDocuments < VdsTest
@@ -8,7 +7,6 @@ class LargeDocuments < VdsTest
     @valgrind = false
     set_owner("vekterli")
 
-    @feed_file = dirs.tmpdir + "feed.json"
     deploy_app(default_app)
     set_expected_logged(//, :slow_processing => true)
     start
@@ -26,42 +24,8 @@ class LargeDocuments < VdsTest
      count = 5
      size = 50000000
 
-     create_dummy_feed(count, size)
-     assert(check_dummy_feed(count, size))
-  end
-
-  def create_dummy_feed(count, size)
-    docs = DocumentSet.new
-    count.times { |id|
-      doc = Document.new("music", "id:test:music:n=" + id.to_s + ":1")
-      content = ""
-      (size / 50).times {
-         content << "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\n"
-      }
-      doc.add_field("bodyfield", content)
-      docs.add(doc)
-    }
-    docs.write_json(@feed_file)
-    feed(:file => @feed_file)
-  end
-
-  def check_dummy_feed(count, size)
-    count.times { |id|
-      documentid="id:test:music:n=" + id.to_s + ":1"
-      target_node.execute("vespa-get " + documentid + " >#{Environment.instance.vespa_home}/tmp/gettmp")
-
-      filesize = File.size("#{Environment.instance.vespa_home}/tmp/gettmp")
-      File.delete("#{Environment.instance.vespa_home}/tmp/gettmp")
-
-      return false if filesize < size
-    }
-
-    target_node.execute("vespa-visit --maxpending 1 --maxpendingsuperbuckets 1 --maxbuckets 1 >#{Environment.instance.vespa_home}/tmp/visittmp")
-    filesize = File.size("#{Environment.instance.vespa_home}/tmp/visittmp")
-
-    File.delete("#{Environment.instance.vespa_home}/tmp/visittmp")
-
-    filesize >= (size * count)
+     target_node.create_dummy_feed(count, size)
+     target_node.check_dummy_feed(count, size)
   end
 
   def teardown
